+++ chirp/src/loader/internal.rs
@@ -1,11 +1,11 @@
 use std::marker::PhantomData;
 
-use bevy::asset::{LoadContext, LoadedAsset};
+use bevy::asset::LoadContext;
 use bevy::ecs::prelude::*;
-use bevy::reflect::TypeRegistryInternal as TypeRegistry;
+use bevy::reflect::TypeRegistry;
 use bevy::scene::Scene;
 
-use super::spawn::Chirp;
+use super::spawn::Chirp_;
 use crate::{interpret, ChirpReader, Handles, ParseDsl};
 
 pub(super) struct Loader<'a, 'r, 'w, 'h, D> {
@@ -20,19 +20,16 @@ impl<'a, 'r, 'w, 'h, D: ParseDsl + 'static> Loader<'a, 'r, 'w, 'h, D> {
         Self { ctx, registry: reg, handles: h, _dsl: PhantomData }
     }
 
-    pub(super) fn load(&mut self, file: &[u8]) {
-        let load = LoadedAsset::new;
-
-        let chirp = match self.load_scene(file) {
+    pub(super) fn load(&mut self, file: &[u8]) -> Chirp_ {
+        match self.load_scene(file) {
             Ok((root, scene)) => {
-                Chirp::Loaded(root, self.ctx.set_labeled_asset("Scene", load(scene)))
+                Chirp_::Loaded(root, self.ctx.add_labeled_asset("Scene".to_owned(), scene))
             }
             Err(errors) => {
                 log_miette_error!(&errors);
-                Chirp::Error(errors)
+                Chirp_::Error(errors)
             }
-        };
-        self.ctx.set_default_asset(LoadedAsset::new(chirp));
+        }
     }
     fn load_scene(&mut self, file: &[u8]) -> Result<(Entity, Scene), interpret::Errors> {
         let mut world = World::new();