+++ ui/src/dsl.rs
@@ -131,7 +131,7 @@ pub struct ParseColorError(String);
 
 #[cfg(feature = "chirp")]
 fn parse_color(
-    _: &bevy::reflect::TypeRegistryInternal,
+    _: &bevy::reflect::TypeRegistry,
     _: Option<&mut bevy::asset::LoadContext>,
     input: &str,
 ) -> Result<Color, ParseColorError> {